fix: resolve all cppcheck defects to pass CI
All 35 defects reported by cppcheck (CI's 'pio check --fail-on-defect
low/medium/high') are now cleared across 10 files. Changes by category:

Real bug fixes:
- RemindersRenderer.cpp: cast uint8_t snprintf args to unsigned (%u mismatch)
- SettingsActivity.cpp: const auto& in range-for (was non-const, read-only loop)
- FileBrowserActivity.cpp: const std::string& for getFileExtension param
- CrossPointWebServer.cpp: const ref for upload in handleFontUploadData
- JsonSettingsIO.cpp: remove duplicate doc["language"] assignment

Inline suppressions for false positives / intentional patterns:
- badBitmaskCheck (16 occurrences): ArduinoJson operator| used as default-value
  API, not bitwise OR. Affects RemindersState.cpp, GoogleClient.cpp,
  JsonSettingsIO.cpp, FontDownloadActivity.cpp.
- useStlAlgorithm (6 occurrences): early-return and accumulation loops in
  FontDownloadActivity.cpp where std::any_of/accumulate would be less clear.
- cstyleCast (3 occurrences): byte-offset pointer arithmetic on struct base in
  JsonSettingsIO.cpp (existing pattern, reinterpret_cast offers no safety gain).
- variableScope: static lastPowerTapMs in main.cpp loop (intentional persistence).
- unreadVariable: SettingsList.h lambda-local vector (cppcheck false positive
  on returned variables inside immediately-invoked lambdas).
- shadowVariable: HTTPUpload& upload in CrossPointWebServer.cpp upload handlers.

Verified: cppcheck 2.17.1 with CI's exact flags (--enable=all, inline-suppr,
--suppress=unusedFunction/missingIncludeSystem/unmatchedSuppression) reports
zero defects across all of src/.

+++ src/activities/reminders/RemindersRenderer.cpp
@@ -298,7 +298,8 @@ uint8_t RemindersRenderer::drawLayout(GfxRenderer& renderer, const RemindersData
   if (startIndex == 0 && !hasMore) {
     snprintf(info, sizeof(info), "%s %u", tr(STR_REMINDERS_TOTAL), data.count);
   } else {
-    snprintf(info, sizeof(info), "%u-%u / %u", startIndex + 1, nextIndex, data.count);
+    snprintf(info, sizeof(info), "%u-%u / %u", static_cast<unsigned>(startIndex + 1), static_cast<unsigned>(nextIndex),
+             static_cast<unsigned>(data.count));
   }
   renderer.drawText(DETAIL_FONT, contentLeft, infoTop, info, true, EpdFontFamily::BOLD);
